Put a lot of stuff in the tup generator module
