+++ hatch/src/constants.rs
@@ -19,4 +19,5 @@ pub static UPDATE_NAME: &str = "update";
 pub static RUN_NAME: &str = "run";
 pub static TEST_NAME: &str = "test";
 
-pub static CATCH_HEADER_URL: &str = "https://github.com/catchorg/Catch2/releases/download/v2.1.1/catch.hpp";
+pub static CATCH_HEADER_URL: &str =
+    "https://github.com/catchorg/Catch2/releases/download/v2.1.1/catch.hpp";
