+++ hatch/src/generators/tup/catch_definition.rs
@@ -1,0 +1,3 @@
+pub fn make_catch_definition_string() -> String {
+    String::from("#define CATCH_CONFIG_MAIN\n#include \"catch.hpp\"")
+}
