+++ hatch/src/generators/tup/catch_header.rs
