+++ hatch/src/generators/tup/test_tupfile.rs
@@ -1,0 +1,3 @@
+pub fn make_test_tupfile_string() -> String {
+    String::from(".gitignore")
+}
