+++ hatch/src/project/target.rs
@@ -1,2 +1,5 @@
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub enum Target { Debug, Release }
+pub enum Target {
+    Debug,
+    Release,
+}
